+++ src/Webcrawler/TwitterWebcrawlerVerifier.php
@@ -16,6 +16,6 @@ class TwitterWebcrawlerVerifier implements WebcrawlerVerifierInterface
      */
     public function verify($ip)
     {
-        return Range::inRange($this->allowedRanges, $ip);
+        return Range::inRange($ip, $this->allowedRanges);
     }
 }